[Update]トップページ

+++ app/controllers/customers/orders_controller.rb
@@ -19,23 +19,29 @@ class Customers::OrdersController < ApplicationController
   	 #注文テーブル保存
   	  order = Order.new(order_params)
       order.total_price = params[:total_price]
-      order.save!
-     #注文詳細テーブル保存
-    current_customer.carts.each do |n|
-      order_detail = OrderDetail.new
-       	order_detail.quantity = n.quantity
-       	order_detail.item_id = n.item_id
-      order_detail.tax = (order.total_price * 0.1).to_i
-      order_detail.order_id = order.id
-      order_detail.save!
-      #在庫計算
-      order_detail.item.stock -= order_detail.quantity
-      order_detail.item.update(stock: order_detail.item.stock)
-  	end
-      #カート削除
-      current_customer.carts.where(customer_id:current_customer.id).destroy_all
-
-      redirect_to customers_homes_thanks_path
+      if order.save
+        #注文詳細テーブル保存
+        current_customer.carts.each do |n|
+          order_detail = OrderDetail.new
+           	order_detail.quantity = n.quantity
+           	order_detail.item_id = n.item_id
+          order_detail.tax = (order.total_price * 0.1).to_i
+          order_detail.order_id = order.id
+          order_detail.save!
+          #在庫計算
+          order_detail.item.stock -= order_detail.quantity
+          order_detail.item.update(stock: order_detail.item.stock)
+  	     end
+        #カート削除
+        current_customer.carts.where(customer_id:current_customer.id).destroy_all
+        redirect_to customers_homes_thanks_path
+      else
+        @order = Order.new
+        @customer = current_customer
+        @carts = Cart.where(customer_id: @customer.id)
+        @addresses = Address.where(customer_id: @customer.id)
+        render :new
+      end
   end
 
   def pay
